fix: Extend saved ugc id to 64 bits (#1885)
Tetsed that ugc models are saved and loaded with the correct bits

+++ dDatabase/GameDatabase/ITables/IPropertyContents.h
@@ -23,7 +23,7 @@ public:
 	// Inserts a new UGC model into the database.
 	virtual void InsertNewUgcModel(
 		std::stringstream& sd0Data,
-		const uint32_t blueprintId,
+		const uint64_t blueprintId,
 		const uint32_t accountId,
 		const LWOOBJID characterId) = 0;
